FEAT: Added Request template
This is template for work order request from a specific (Sysaid) ticketing system. I primarily used it in a specific role but it is complex enough that it can be a good example to reference in the future. I also used org-journal with this template, so be sure to install that or modify the template to not use org-journal.